+++ packages/golden-layout/src/index.ts
@@ -7,5 +7,3 @@ export * from './controls';
 export * from './errors';
 export * from './items';
 export * from './utils';
-export { default as GoldenLayoutThemeExport } from './GoldenLayoutThemeExport';
-export type { DragSourceEvent } from './LayoutManager';